+++ data/cube/aggregate/LeafCountAggregator.ts
@@ -20,4 +20,12 @@ export class LeafCountAggregator extends Aggregator {
     override replace(rows, currAgg, update, context) {
         return currAgg;
     }
+
+    override add(rows, currAgg, value, context) {
+        return currAgg + 1;
+    }
+
+    override remove(rows, currAgg, value, context) {
+        return currAgg - 1;
+    }
 }